Better name for database 'driver'
